refactor: extract a runtime schema package
