+++ packages/orm/src/schema.ts
@@ -1,0 +1,2 @@
+export * from '@zenstackhq/schema';
+export type { OperandExpression } from 'kysely';
