--- packages/orm/src/schema/index.ts
@@ -1,4 +1,0 @@
-export type * from '@zenstackhq/sdk/schema';
-export type { OperandExpression } from 'kysely';
-export * from './auth';
-export * from './expression';
