+++ packages/schema/package.json
@@ -1,7 +1,7 @@
 {
-    "name": "@zenstackhq/sdk",
+    "name": "@zenstackhq/schema",
     "version": "3.0.0-beta.14",
-    "description": "ZenStack SDK",
+    "description": "ZenStack Runtime Schema",
     "type": "module",
     "scripts": {
         "build": "tsc --noEmit && tsup-node",
@@ -28,16 +28,10 @@
         }
     },
     "dependencies": {
-        "@zenstackhq/language": "workspace:*",
-        "@zenstackhq/common-helpers": "workspace:*",
-        "langium": "catalog:",
-        "ts-pattern": "catalog:",
-        "typescript": "catalog:"
+        "decimal.js": "catalog:"
     },
     "devDependencies": {
         "@zenstackhq/eslint-config": "workspace:*",
-        "@zenstackhq/typescript-config": "workspace:*",
-        "decimal.js": "catalog:",
-        "kysely": "catalog:"
+        "@zenstackhq/typescript-config": "workspace:*"
     }
 }
