+++ packages/schema/src/index.ts
@@ -1,4 +1,3 @@
 export type * from './expression';
+export * from './expression-utils';
 export type * from './schema';
-
-export type { OperandExpression } from 'kysely';
