+++ tests/e2e/orm/schemas/basic/helper.ts
@@ -1,4 +1,4 @@
-import type { DataSourceProviderType } from '@zenstackhq/sdk/schema';
+import type { DataSourceProviderType } from '@zenstackhq/schema';
 import { schema } from './schema';
 
 export function getSchema<ProviderType extends DataSourceProviderType>(type: ProviderType) {
