+++ tests/e2e/package.json
@@ -15,6 +15,7 @@
         "@paralleldrive/cuid2": "^2.2.2",
         "@zenstackhq/cli": "workspace:*",
         "@zenstackhq/language": "workspace:*",
+        "@zenstackhq/schema": "workspace:*",
         "@zenstackhq/orm": "workspace:*",
         "@zenstackhq/plugin-policy": "workspace:*",
         "@zenstackhq/sdk": "workspace:*",
